backend/openai-agents-py/pyproject.toml
[tool.poetry]
name = "travel-planner-openai-agents"
version = "0.1.0"
description = "Travel planning system using OpenAI Agents SDK"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "app"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"